requirements.txt
Flask
gunicorn
tflite-runtime          # The lightweight TensorFlow Lite interpreter
opencv-python-headless  # Server version of OpenCV
mtcnn
google-generativeai
Pillow
numpy                   # mtcnn might need this explicitly